requirements.txt
Flask
google-generativeai
yagmail
docxtpl
openpyxl
gunicorn
requests